// backend/src/common/enums/task-permission.enum.ts
export enum TaskPermission {
  COMPANY_MANAGEMENT = 'COMPANY_MANAGEMENT',
  BRANCH_MANAGEMENT = 'BRANCH_MANAGEMENT',
  USER_MANAGEMENT = 'USER_MANAGEMENT',
  DESIGN_ENTRIES = 'DESIGN_ENTRIES',
  ORDER_ENTRIES = 'ORDER_ENTRIES',
  ORDER_APPROVALS = 'ORDER_APPROVALS',
  PRICING_CONFIGURATION = 'PRICING_CONFIGURATION',
  VIEW_REPORTS = 'VIEW_REPORTS',
}